+++ protocol/test/foundry/Sun.t.sol
@@ -130,19 +130,19 @@ contract SunTest is Sun, Test {
   function test_deltaB_positive_podRate_low() public {
     field.incrementTotalPodsE(100);
     season.sunSunrise(300e6, 0); // deltaB = +300; case 0 = low pod rate
-    assertEq(field.totalSoil(), 148); // FIXME: how calculated?
+    assertEq(uint256(field.totalSoil()), 148); // FIXME: how calculated?
   }
   
   function test_deltaB_positive_podRate_medium() public {
     field.incrementTotalPodsE(100);
     season.sunSunrise(300e6, 8); // deltaB = +300; case 0 = medium pod rate
-    assertEq(field.totalSoil(), 99); // FIXME: how calculated?
+    assertEq(uint256(field.totalSoil()), 99); // FIXME: how calculated?
   }
 
   function test_deltaB_positive_podRate_high() public {
     field.incrementTotalPodsE(100);
     season.sunSunrise(300e6, 8); // deltaB = +300; case 0 = high pod rate
-    assertEq(field.totalSoil(), 99); // FIXME: how calculated?
+    assertEq(uint256(field.totalSoil()), 99); // FIXME: how calculated?
   }
 
   ///////////////////////// Minting /////////////////////////
@@ -204,19 +204,19 @@ contract SunTest is Sun, Test {
   //   // low pod rate
   //   field.incrementTotalPodsE(100);
   //   season.sunSunrise(300e6, 0); // deltaB = +300; case 0 = low pod rate
-  //   assertEq(field.totalSoil(), 148); // FIXME: how calculated?
+  //   assertEq(uint256(field.totalSoil()), 148); // FIXME: how calculated?
   //   snapId = _reset(snapId);
 
   //   // medium pod rate
   //   field.incrementTotalPodsE(100);
   //   season.sunSunrise(300e6, 8); // deltaB = +300; case 0 = low pod rate
-  //   assertEq(field.totalSoil(), 99); // FIXME: how calculated?
+  //   assertEq(uint256(field.totalSoil()), 99); // FIXME: how calculated?
   //   snapId = _reset(snapId);
 
   //   // high pod rate
   //   field.incrementTotalPodsE(100);
   //   season.sunSunrise(300e6, 8); // deltaB = +300; case 0 = low pod rate
-  //   assertEq(field.totalSoil(), 99); // FIXME: how calculated?
+  //   assertEq(uint256(field.totalSoil()), 99); // FIXME: how calculated?
   // }
 
 }
